requirements.txt
streamlit>=1.33
pandas
gspread
google-auth
matplotlib
